rename app -> commitStatus
